+++ src/app/components/StatusBar/index.tsx
@@ -141,7 +141,7 @@ class StatusBar extends React.Component<IProps, IState> {
     private tryAgainHandler = (e: any) => {
         e.stopPropagation();
         e.preventDefault();
-        APIManager.getInstance().checkNetwork();
+        APIManager.getInstance().checkNetwork(true);
     }
 
     private updateFunctionHandler = () => {